requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# LLM API clients
openai==1.12.0
anthropic==0.18.1

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
